perf: CLIP 画像エンコーダを fp16 化（実機 ANE で高速化）
実機ログで画像埋め込みが 1枚 0.5〜1.3s と遅い。原因は画像エンコーダを
compute_precision=FLOAT32 で変換しており、ANE 非対応で GPU/CPU フォール
バックしていたこと（元はシミュレータの NaN 回避目的）。シミュレータ最適化は
本末転倒のため fp16 のみへ切替。

- convert_mobileclip.py: 画像エンコーダを FLOAT16 に
- fp16 はシミュレータで NaN 化し得るが、ランタイムの有限性チェックが nil に
  落として安全に無効化。画像タワー依存テストはシミュレータでスキップし実機検証へ
- ImageRecognitionTests: @testable import MobileCLIPKit（潜在的な未importを修正）
  ＋ skipImageTowerOnSimulator ガード。テキスト/翻訳系は常時実行
- CLAUDE.md / 設計資料 / records(ADR-11・事例) を更新

※ 反映には bash scripts/build_mobileclip.sh でのモデル再生成＋実機再配置が必要。

+++ scripts/convert_mobileclip.py
@@ -64,9 +64,12 @@ img_ml = ct.convert(
     inputs=[ct.ImageType(name="image", shape=img_example.shape, scale=1 / 255.0, bias=[0, 0, 0])],
     outputs=[ct.TensorType(name="embedding")],
     minimum_deployment_target=ct.target.iOS17,
-    # ⚠️ 既定の FLOAT16 だと画像タワー（conv/transformer ハイブリッド）がシミュレータで
-    # 数値オーバーフローし、埋め込みが全 NaN になる。FLOAT32 で変換して回避する。
-    compute_precision=ct.precision.FLOAT32,
+    # 画像エンコーダは FLOAT16 で変換する。実機の Neural Engine(ANE) は fp16 前提のため、
+    # fp32 だと ANE に載らず GPU/CPU フォールバックで埋め込みが大幅に遅くなる（実機ログで確認・
+    # 8枚あたり 4〜10s）。fp16 化で ANE 実行＝高速化を狙う。
+    # ※ fp16 は一部シミュレータで数値オーバーフローし埋め込みが NaN になり得るが、ランタイムの
+    #   有限性チェックが nil に落として安全に無効化するだけ（CLIP 機能の検証・本番は実機で行う）。
+    compute_precision=ct.precision.FLOAT16,
 )
 img_ml.save(os.path.join(OUT, "MobileCLIPImageS2.mlpackage"))
 print("saved image encoder")
